+++ js/validate-form.js
@@ -1,0 +1,49 @@
+import { hashtagInput } from './picture-upload.js';
+
+const hashtagRegExp = /^#[a-zа-яё0-9]{1,19}$/i;
+let errorHashtagMessage = '';
+
+
+function validateComment(value) {
+  return value.length <= 140;
+}
+
+
+function validateHashtags() {
+  const hashtags = hashtagInput.value.trim().split(/\s+/);
+  const uniqueHashtags = new Set(hashtags.map((item) => item.toLowerCase()));
+  let isValid = true;
+
+  if (hashtags.length === 0) {
+    return true;
+  }
+
+  if (uniqueHashtags.size > 5) {
+    isValid = false;
+    errorHashtagMessage = 'Нельзя указать больше пяти хэш-тегов';
+  } else {
+    for (const hashtag of hashtags) {
+      if (hashtag === '') {
+        continue;
+      }
+      if (!hashtagRegExp.test(hashtag)) {
+        isValid = false;
+        errorHashtagMessage = `Хэш-тег "${hashtag}" некорректен`;
+        break;
+      }
+      if (hashtags.indexOf(hashtag) !== hashtags.lastIndexOf(hashtag)) {
+        isValid = false;
+        errorHashtagMessage = `Один и тот же хэш-тег "${hashtag}" не может быть использован дважды`;
+        break;
+      }
+    }
+  }
+
+  return isValid;
+}
+
+export {
+  errorHashtagMessage,
+  validateComment,
+  validateHashtags
+};
